pyproject.toml
[tool.poetry]
name = ""
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
django-environ = "^0.9.0"
pytest = "^7.2.0"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
mypy = "^0.991"
isort = "^5.10.1"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
